+++ CHANGELOG.md
@@ -1,2 +1,11 @@
+## v0.0.3 / 2016-03-30
+- Added route wildcards :any, :num, :word
+- Parse query string
+- Set Interval as private property
+- Fixed remove route
+
 ## v0.0.2 / 2016-03-30
 - Fix URI match issue
+
+## v0.0.1 / 2016-03-30
+- Aded core functionality